refactoring after tests it's almost like TDD Extracted the class BulkPackageSearch and made the test on *that* as the test is simpler and clearer
