+++ NuKeeper/NuGet/Api/PackageUpdatesLookup.cs
@@ -1,24 +1,27 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
-using NuGet.Protocol.Core.Types;
 using NuKeeper.RepositoryInspection;
 
 namespace NuKeeper.NuGet.Api
 {
     public class PackageUpdatesLookup : IPackageUpdatesLookup
     {
-        private readonly IApiPackageLookup _packageLookup;
+        private readonly BulkPackageSearch _bulkSearch;
 
         public PackageUpdatesLookup(IApiPackageLookup packageLookup)
         {
-            _packageLookup = packageLookup;
+            _bulkSearch = new BulkPackageSearch(packageLookup);
         }
 
         public async Task<List<PackageUpdateSet>> FindUpdatesForPackages(IEnumerable<PackageInProject> packages)
         {
-            var latestVersions = await BuildLatestVersionsDictionary(packages);
+            var packageIds = packages
+                .Select(p => p.Id)
+                .Distinct();
+
+            var latestVersions = await _bulkSearch.LatestVersions(packageIds);
+
             var results = new List<PackageUpdateSet>();
 
             foreach (var packageId in latestVersions.Keys)
@@ -37,34 +40,6 @@ namespace NuKeeper.NuGet.Api
             }
 
             return results;
-        }
-
-        private async Task<Dictionary<string, IPackageSearchMetadata>> BuildLatestVersionsDictionary(IEnumerable<PackageInProject> packages)
-        {
-            var packageIds = packages
-                .Select(p => p.Id)
-                .Distinct();
-
-            var lookupTasks = packageIds
-                .Select(id => _packageLookup.LookupLatest(id))
-                .ToList();
-
-            await Task.WhenAll(lookupTasks);
-
-            var result = new Dictionary<string, IPackageSearchMetadata>();
-
-            foreach (var lookupTask in lookupTasks)
-            {
-                var serverVersion = lookupTask.Result;
-                if (serverVersion?.Identity != null)
-                {
-                    var packageId = serverVersion.Identity.Id;
-                    result.Add(packageId, serverVersion);
-                    Console.WriteLine($"Found latest version of {packageId}: {serverVersion.Identity.Id} {serverVersion.Identity.Version}");
-                }
-            }
-
-            return result;
         }
     }
 }
